@@ -1,7 +1,7 @@
 """ API Blueprint Application """
 
 from flask import Blueprint, current_app
-from flask_restplus import Api
+from flask_restx import Api
 
 api_bp = Blueprint('api_bp', __name__, url_prefix='/api')
 api_rest = Api(api_bp)
