@@ -5,7 +5,7 @@ http://flask-restplus.readthedocs.io
 
 from datetime import datetime
 from flask import request
-from flask_restplus import Resource
+from flask_restx import Resource
 
 from .security import require_auth
 from . import api_rest
