@@ -1,21 +1,21 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
 import Home from './views/Home.vue'
 import Api from './views/Api.vue'
 
-Vue.use(Router)
+const routes = [
+  {
+    path: '/',
+    name: 'home',
+    component: Home
+  },
+  {
+    path: '/api',
+    name: 'api',
+    component: Api
+  }
+]
 
-export default new Router({
-  routes: [
-    {
-      path: '/',
-      name: 'home',
-      component: Home
-    },
-    {
-      path: '/api',
-      name: 'api',
-      component: Api
-    }
-  ]
+export default createRouter({
+  history: createWebHistory(),
+  routes
 })
